refactor: map borehole.organisation and borehole.codingStandard
The attribute 'borehole.organisation' is mandatory, but no information are
available in source data. So for all attributes in organisation nilReason =
'witheld' is assigned, expect organisation.role. Here 'resourceProvider' is
assigned, like required in the boreholeML documentation.

'boreholes.swissgeol.ch_v01' is assigned to borehole.codingStandard.

Also the 'assign' to 'borehole.totalLength.uom' is changed to a 'assign bound'.

SVC-1486