fix: correct mapping of attributes in 'borehole.organisation'
Previously, a nilReason was assigned to 'contactInfo', 'individualName',
'organisationName'and 'positionName'. But this is not allowed according
the schema.
The BoreholeML documentation says that only the attributes 'organisationName',
'contactInfo.address.country' and 'administrativeArea' are mandatory. As these
attributes are not present in the source data the value 'unbekannt' is assigned
for now.
The respective changes are done here.

SVC-1486